requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
Jinja2>=3.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
Flask-Compress>=1.13.0
